Add MIME mappings for fonts
This ensures that the font files which are not mapped by default on older
versions of IIS will be mapped, so that IIS will serve them correctly
(otherwise it will return a 404).  This list is based on the default
mappings in IIS 7 (from https://support.microsoft.com/en-us/help/936496)
and the MIME types are the defaults from an IIS 10 installation.

These additions will override any system-wide defaults set (though they
should match the defaults present in newer servers).  These instructions
will _not_ overwrite any existing mappings in the `web.config` file for
the specific site into which the theme is installed (i.e. if `.woff` has
already been mapped in the site, this won't change that mapping, but if
`.woff2` has not been mapped, that mapping will be added).  Any other
customizations to the `staticContent` section of the `web.config` file
will also be preserved.

Fixes #60