Update MainMenu.txt for slight better "target" (#244)
* Update MainMenu.txt for slight better "target"

Added the conditional wrapper on each occurrence of target=, this way, for the majority of links that are local and do not open in a new window, you don't end up with menus full of empty target attributes. (e.g. `target=""`).
 
Warning: I did test this. It also looks right. But your mileage may vary. :)

* Update menus/main/MainMenu.txt

Updating to reflect recent changes in color class naming

* Update menus/main/MainMenu.txt

Updating to reflect recent changes in color class naming

Co-authored-by: David Poindexter <dpoindexter@nvisionative.com>